refactor: storage cmds
